Renamed config to configs
