+++ configs/mining.yaml
