+++ pkg/mining/configs/domria.go
@@ -1,4 +1,4 @@
-package config
+package configs
 
 import (
 	"fmt"
