+++ pkg/mining/configs/mining.go
@@ -1,4 +1,4 @@
-package config
+package configs
 
 import (
 	"fmt"
@@ -12,7 +12,7 @@ import (
 func NewMining() (*Mining, error) {
 	_, filePath, _, ok := runtime.Caller(0)
 	if !ok {
-		return nil, fmt.Errorf("config: failed to find the caller path")
+		return nil, fmt.Errorf("configs: failed to find the caller path")
 	}
 	file, err := os.Open(
 		filepath.Join(
@@ -22,18 +22,18 @@ func NewMining() (*Mining, error) {
 		),
 	)
 	if err != nil {
-		return nil, fmt.Errorf("config: failed to open the config file, %v", err)
+		return nil, fmt.Errorf("configs: failed to open the config file, %v", err)
 	}
 	bytes, err := ioutil.ReadAll(file)
 	if err != nil {
-		return nil, fmt.Errorf("config: failed to read the config file, %v", err)
+		return nil, fmt.Errorf("configs: failed to read the config file, %v", err)
 	}
 	if err = file.Close(); err != nil {
-		return nil, fmt.Errorf("config: failed to close the config file, %v", err)
+		return nil, fmt.Errorf("configs: failed to close the config file, %v", err)
 	}
 	var mining Mining
 	if err = yaml.Unmarshal(bytes, &mining); err != nil {
-		return nil, fmt.Errorf("config: failed to unmarshal the config file, %v", err)
+		return nil, fmt.Errorf("configs: failed to unmarshal the config file, %v", err)
 	}
 	return &mining, nil
 }
