+++ pkg/mining/configs/prospectors.go
@@ -1,4 +1,4 @@
-package config
+package configs
 
 import (
 	"fmt"
